Move most Post component tests into unit tests, add PropTypes, fix broken test
