+++ test/components/Post/connector.test.js
@@ -1,0 +1,15 @@
+import '../../support'
+import { mockActionResponse } from '../../support/helpers'
+import { HOST } from '../../../src/util/api'
+import * as connector from '../../../src/components/Post/connector'
+
+const state = {}
+const post = {id: 'testpost'}
+
+// nock(HOST).get('/noo/user/me').reply(200, {})
+
+describe('Post Connector', () => {
+  describe('mapStateToProps', () => {
+    // TODO
+  })
+})
